+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+- Bugfix when there are no servicePoints in Paazl Perfect
 
 ## [1.2.1] - 2017-05-05
 ### Changed